Moved param definitions to type file
